pyproject.toml
[tool.poetry]
name = "penta-real-esrgan"
version = "0.1.0"
description = ""
authors = ["supplepentan <supplepentan@gmail.com>"]
readme = "README.md"
packages = [{include = "penta_real_esrgan"}]

[tool.poetry.dependencies]
python = "^3.8"
torch = { url = "https://download.pytorch.org/whl/cu117/torch-1.13.0%2Bcu117-cp38-cp38-win_amd64.whl"}
torchvision = { url = "https://download.pytorch.org/whl/cu117/torchvision-0.14.0%2Bcu117-cp38-cp38-win_amd64.whl"}
basicsr = "^1.4.2"
facexlib = "^0.2.5"
gfpgan = "^1.3.8"
opencv-python = "^4.6.0.66"
tqdm = "^4.64.1"
fastapi = "^0.85.2"
uvicorn = {extras = ["standard"], version = "^0.19.0"}
jinja2 = "^3.1.2"
pillow = "^9.3.0"
python-multipart = "^0.0.5"
black = "^22.10.0"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
